project/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
distro==1.9.0
Django==4.2.10
django-admin==2.0.2
django-environ==0.11.2
django-excel-response2==3.0.6
django-six==1.0.5
djangorestframework==3.14.0
excel-base==1.0.4
exceptiongroup==1.2.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
inflection==0.5.1
isoweek==1.3.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
multidict==6.0.5
openai==1.12.0
openapi==1.1.0
pydantic==2.6.2
pydantic_core==2.16.3
Pygments==2.17.2
PyPDF2==3.0.1
python-dateutil==2.8.2
pytz==2024.1
referencing==0.33.0
requests==2.31.0
rpds-py==0.18.0
screen==1.0.1
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.4
TimeConvert==3.0.13
tqdm==4.66.2
typing_extensions==4.10.0
tzlocal==5.2
urllib3==2.2.1
xlwt==1.3.0
yarl==1.9.4
